+++ api/src/entity/project-locale.entity.ts
@@ -1,8 +1,9 @@
-import { Column, Entity, Index, ManyToOne, PrimaryGeneratedColumn } from 'typeorm';
+import { Column, Entity, Index, ManyToOne, PrimaryGeneratedColumn, OneToMany, JoinColumn } from 'typeorm';
 
 import { AccessTimestamps } from './base';
 import { Locale } from './locale.entity';
 import { Project } from './project.entity';
+import { Translation } from './translation.entity';
 
 @Entity()
 @Index(['project', 'locale'], { unique: true })
@@ -15,6 +16,9 @@ export class ProjectLocale {
 
   @ManyToOne(() => Project, { onDelete: 'CASCADE' })
   project: Project;
+
+  @OneToMany(() => Translation, translation => translation.projectLocale)
+  translations: Translation[];
 
   @Column(type => AccessTimestamps)
   date: AccessTimestamps;