+++ docs/changelog.md
@@ -8,6 +8,7 @@ To download the source code for each release, check out [GitHub](https://github.
 
 ## next
 - Automatically detect file import format.
+- [PR-89:](https://github.com/traduora/traduora/pull/89) Added translation progress stats per locale.
 - Increase max size for file import (1MB).
 - Improve error message when importing files which are too large.